cleanup ingest event
