+++ functions/src/logic/utils/get-document-from-url.ts
